+++ deepl-java/src/test/java/com/deepl/api/TranslationMemoryTest.java
@@ -3,6 +3,7 @@
 // license that can be found in the LICENSE file.
 package com.deepl.api;
 
+import java.io.File;
 import java.util.List;
 import org.junit.jupiter.api.*;
 
@@ -57,5 +58,62 @@ public class TranslationMemoryTest extends TestBase {
                 .setTranslationMemoryThreshold(80));
 
     Assertions.assertNotNull(result);
+  }
+
+  @Test
+  void testTranslateDocumentWithTranslationMemoryId() throws Exception {
+    // Note: this test may use the mock server that will not translate the document
+    // with a translation memory, therefore we do not check the translated result.
+    Assumptions.assumeTrue(isMockServer);
+    DeepLClient client = createDeepLClient();
+    File inputFile = createInputFile("Hallo, Welt!");
+    File outputFile = createOutputFile();
+
+    client.translateDocument(
+        inputFile,
+        outputFile,
+        "de",
+        "en-US",
+        new DocumentTranslationOptions().setTranslationMemoryId(DEFAULT_TM_ID));
+
+    Assertions.assertNotNull(readFromFile(outputFile));
+  }
+
+  @Test
+  void testTranslateDocumentWithTranslationMemoryIdAndThreshold() throws Exception {
+    Assumptions.assumeTrue(isMockServer);
+    DeepLClient client = createDeepLClient();
+    File inputFile = createInputFile("Hallo, Welt!");
+    File outputFile = createOutputFile();
+
+    client.translateDocument(
+        inputFile,
+        outputFile,
+        "de",
+        "en-US",
+        new DocumentTranslationOptions()
+            .setTranslationMemoryId(DEFAULT_TM_ID)
+            .setTranslationMemoryThreshold(80));
+
+    Assertions.assertNotNull(readFromFile(outputFile));
+  }
+
+  @Test
+  void testTranslateDocumentWithTranslationMemoryInfo() throws Exception {
+    Assumptions.assumeTrue(isMockServer);
+    DeepLClient client = createDeepLClient();
+    List<TranslationMemoryInfo> translationMemories = client.listTranslationMemories(0, 10);
+    TranslationMemoryInfo translationMemory = translationMemories.get(0);
+    File inputFile = createInputFile("Hallo, Welt!");
+    File outputFile = createOutputFile();
+
+    client.translateDocument(
+        inputFile,
+        outputFile,
+        "de",
+        "en-US",
+        new DocumentTranslationOptions().setTranslationMemory(translationMemory));
+
+    Assertions.assertNotNull(readFromFile(outputFile));
   }
 }